removed all warnings, finished merge of branches. still have to fix font colors

+++ TotalCrossVM/src/dx/PhoneDirect3DXamlAppComponent.cpp
@@ -43,7 +43,7 @@ void Direct3DBackground::OnPointerPressed(DrawingSurfaceManipulationHost^ sender
 {
 	auto pos = args->CurrentPoint->Position;
 	debug("pressed lastY %.2f lastX %.2f Y %.2f X %.2f", lastY, lastX, pos.Y, pos.X);
-	eventQueuePush(PENEVENT_PEN_DOWN, 0, lastX = pos.X, lastY = pos.Y - glShiftY, -1);
+   eventQueuePush(PENEVENT_PEN_DOWN, 0, (int32)(lastX = pos.X), (int32)(lastY = pos.Y - glShiftY), -1);
 }
 
 void Direct3DBackground::OnPointerMoved(DrawingSurfaceManipulationHost^ sender, PointerEventArgs^ args)
@@ -52,7 +52,7 @@ void Direct3DBackground::OnPointerMoved(DrawingSurfaceManipulationHost^ sender, 
 	auto pos = args->CurrentPoint->Position;
 	if (lastX != pos.X || lastY != pos.Y) {
 		debug("moving lastY %.2f lastX %.2f Y %.2f X %.2f", lastY, lastX, pos.Y, pos.X);
-		eventQueuePush(PENEVENT_PEN_DRAG, 0, lastX = pos.X, lastY = pos.Y - glShiftY, -1);
+      eventQueuePush(PENEVENT_PEN_DRAG, 0, (int32)(lastX = pos.X), (int32)(lastY = pos.Y - glShiftY), -1);
 		isDragging = true;
 	}
 }
@@ -61,7 +61,7 @@ void Direct3DBackground::OnPointerReleased(DrawingSurfaceManipulationHost^ sende
 {
 	// Insert your code here.
 	auto pos = args->CurrentPoint->Position;
-	eventQueuePush(PENEVENT_PEN_UP, 0, lastX = pos.X, lastY = pos.Y - glShiftY, -1);
+   eventQueuePush(PENEVENT_PEN_UP, 0, (int32)(lastX = pos.X), (int32)(lastY = pos.Y - glShiftY), -1);
 	isDragging = false;
 }
